chore: Extract UnionSchema to its own module (src/schema/union.rs)
No functional changes!
